==> 00_designing-patches.Rmd <==
---
title: "Designing patches representing the ecosystems of the Southeast US"
author: Eric R. Scott
output: html_document
date: '2023-01-11'
---

```{r}
library(spatstat.data)
library(spatstat.geom)
library(tidyverse)
library(ggforce) #for geom_circle
```

# Ecoregions

[Southern Coastal Plain](https://bplant.org/region/134):

-   longleaf pine flatwoods and savannas

-   richer forests with slash pine, pond pine, pond cypres, american sweetgum, southern magnolia, laurel oak, white oak, american beech (i.e. mix of evergreen, evergreen hardwood, and deciduous hardwood PFTs)

-   Floodplains with bald cypres, pond cypres, water tupelo, sweetgum, green ash, water hickory ("temperate.Hydric" PFT??)

-   Florida scrub

-   monoculture pine plantations

[Southeastern plains](https://bplant.org/region/123):

-   more hardwoods and more closed-canopy than southern coastal plain

-   pine forests with longleaf pine and loblolly pine

-   Mixed oak-hickory-pine forests

-   southeastern mixed forest with mix of braodleaf evergreens, deciduous evergreens (SIC), and pines

-   Floodplains had deciduous forests

-   Cypress swamps

[Piedmont](https://bplant.org/region/102):

-   Different bedrock / soil than other regions

-   oak-pine-hickory forest with various Carya species, shortleaf pine, loblolly pine, white oak, post oak

-   Agriculture, old-field, and early successional forest

Can probably summarize this as:

1.  pine forest

2.  oak-hickory-pine forest (deciduous hardwood and pine)

3.  mixed forest (pine, deciduous hardwood, evergreen hardwood)

4.  prairie (c4 and c3 grass, maybe forb)

5.  old field (like above but with more forbs?) or early successional forest (like prairie but with small trees)

## Pine Forest

There's a longleaf pine dataset in the `spatstat.data` package that we can use.
It's 584 Longleaf pines in a 200m x 200m region in southern Georgia.
Data collected by Platt, Evans and Rathbun (1988).

```{r}
plot(longleaf)
longleaf_df <- as.data.frame(longleaf) |> 
  rename(dbh = marks)

ggplot(longleaf_df) +
  geom_circle(aes(x0 = x, y0 = y, r = dbh/2/10)) +
  coord_equal()
```

Data needs:

-   number of cohorts (bins)
-   dbh for each cohort
-   density (stem/m\^2) for each cohort

```{r}
ggplot(longleaf_df, aes(dbh)) + geom_histogram()
```

Looks like we could use 2 or 3 cohorts:

Saplings (dbh 1--20 ish), mature trees (dbh 20-50 ish), old trees (50-70 ish)

### PFTs

-   temperate.Southern_Pine (7)
-   SetariaWT (1)

```{r}
pine <- 
  longleaf_df |> 
  mutate(group = case_when(
    dbh < 20 ~ "sapling",
    dbh >=20 & dbh < 50 ~ "mature tree",
    dbh >=50 ~ "old growth"
  )) |> 
  group_by(group) |> 
  summarize(
    n = n(),
    median_dbh = median(dbh),
    density = n/(200*200),
  ) |> 
  mutate(pft = 7) |> 
  add_row(group = "setaria", n = NA, median_dbh = 1, density = 1, pft = 1)
pine
```



## Mixed forest

Deciduous hardwood, evergreen hardwood, conifer

### PFTs

-   temperate.South_Mid_Hardwood (10)
-   temperate.Southern_Pine (7)
-   temperate.Evergreen_Hardwood (8)
-   SetariaWT (1)

```{r}
mixed <- tribble(
  ~cohort, ~dbh, ~ density, ~pft,
  "mature_hardwood", 40, 0.004, 10,
  "sapling_hardwood", 7, 0.004, 10,
  "mature_pine", 40, 0.004, 7,
  "sapling_pine", 7, 0.004, 7,
  "mature_evergreen", 40, 0.001, 8,
  "sapling_evergreen", 7, 0.001, 8,
  "setaria", 1, 1, 1
)
mixed
```

Try to visualize to see if it makes sense.  This isn't perfect because trees can end up on top of each other depending on the seed.

```{r}
set.seed(123)

mixed |> 
  filter(cohort != "setaria") |> 
  mutate(m2 = 10000) |>
  mutate(n = m2 * density) |> 
  group_by(cohort) |> 
  mutate(dbh_rand = list(rnorm(n = n, mean = dbh, sd = 5))) |> 
  unnest(dbh_rand) |> 
  ungroup() |> 
  mutate(x = runif(n = n, 0, sqrt(m2)),
         y = runif(n = n, 0, sqrt(m2))) |> 
  mutate(r_m = dbh_rand/10/2) |> #get radius in meters from diameter in cm
  ggplot() +
  geom_circle(aes(x0 = x, y0 = y, r = r_m, color = cohort)) +
  coord_equal()
```

Seems a little crowded, but not unreasonable.


## Prairie

### PFTs

-   ebifarm.c3grass (5)
-   ebifarm.c4grass (16) 
-   ebifarm.forb (12) 
-   SetariaWT (1)

ebifarm.c4grass and ebifarm.forb might cause issues because they are using weird ED2 PFTs to not overlap with the other PFTs

```{r}
prairie <- tribble(
  ~cohort, ~dbh, ~density, ~pft,
  "c4", 1, 3, 16,
  "c3", 1, 3, 5,
  "forb", 1, 0.15, 12,  #density should be like 1/20 of the grasses combined?
  "setaria", 1, 1, 1
)
prairie
```

```{r}
prairie |> 
mutate(m2 = 100) |> 
  mutate(n = m2 * density) |> 
  group_by(cohort) |> 
  mutate(dbh_rand = list(rnorm(n = n, mean = dbh, sd = 0.1))) |> 
  unnest(dbh_rand) |> 
  ungroup() |> 
  mutate(x = runif(n = n, 0, sqrt(m2)),
         y = runif(n = n, 0, sqrt(m2))) |> 
  mutate(r_m = dbh_rand/10/2) |> #get radius in meters from diameter in cm
  ggplot() +
  geom_circle(aes(x0 = x, y0 = y, r = r_m, color = cohort)) +
  coord_equal()
```


# Patch Files

We just need a single pss file because nothing really needs to differ in it between the different ecosystems.

Followed [wiki for IED_INIT_MODE=6](https://github.com/EDmodel/ED2/wiki/Initial-conditions#files-types-and-formats-for-nlied_init_mode6) to create this:

```{r}
pss <- 
  tibble(
  time = 1997, #dummy
  patch = c(1, 2), #start with two patches
  trk = 1, 
  age = c(5, 50),
  area = c(.5, .5), #equal area for all three patches
  water = 0, #dummy
  fsc = 1, #fast soil carbon in Kg/m^2
  stsc = 5, #structural soil carbon
  stsl = 5, #structural soil lignin
  ssc = 0.01, #slow soil carbon
  lai = 0, #dummy
  msn = 1, #mineralized soil nitrogen in Kg/m^2
  fsn = 1, #fast soil nitrogen
  nep = 0, #dummy
  gpp = 0, #dummy
  rh  = 0, #dummy
)
pss
```

# Cohort files

```{r}
pine_css <- 
  pine |> 
  transmute(
    time = 1997, #dummy
    dbh = median_dbh, #diameter at breast height
    hite = 0, #dummy
    pft = pft,
    n = density,
    bdead = 0, #dummy
    balive = 0, #dummy
    lai = 0 #dummy
    )

#make 2 identical patches to initialize run
pine_css <-
  bind_rows(pine_css, pine_css, .id = "patch") |> 
  mutate(cohort = 1:n(), .before = dbh) |> 
  relocate(patch, .before = cohort)
pine_css
```


```{r}
mixed_css <-
  mixed |> 
  transmute(
    time = 1997, #dummy
    dbh = dbh, #diameter at breast height
    hite = 0, #dummy
    pft = pft,
    n = density,
    bdead = 0, #dummy
    balive = 0, #dummy
    lai = 0 #dummy
    )

#make 2 identical patches to initialize run
mixed_css <-
  bind_rows(mixed_css, mixed_css, .id = "patch") |> 
  mutate(cohort = 1:n(), .before = dbh) |> 
  relocate(patch, .before = cohort)
mixed_css
```

```{r}
prairie_css <-
  prairie |> 
  transmute(
    time = 1997, #dummy
    dbh = dbh, #diameter at breast height
    hite = 0, #dummy
    pft = pft,
    n = density,
    bdead = 0, #dummy
    balive = 0, #dummy
    lai = 0 #dummy
    )

#make 2 identical patches to initialize run
prairie_css <-
  bind_rows(prairie_css, prairie_css, .id = "patch") |> 
  mutate(cohort = 1:n(), .before = dbh) |> 
  relocate(patch, .before = cohort)
prairie_css

```


# Write files

Writes them to /data/ shared folder. The 02-setup-runs script will copy these and re-name them appropriately for each run.

```{r}
write.table(pss, "/data/sites/generic_patches/generic.pss", quote = FALSE, row.names = FALSE)
write.table(pine_css, "/data/sites/generic_patches/pine.css", quote = FALSE, row.names = FALSE)
write.table(mixed_css, "/data/sites/generic_patches/mixed.css", quote = FALSE, row.names = FALSE)
write.table(prairie_css, "/data/sites/generic_patches/prairie.css", quote = FALSE, row.names = FALSE)
```
