==> 01_wrangle-site-data.Rmd <==
---
title: "Create site data"
author: "Eric R. Scott"
date: "2022-10-13"
output: html_document
---

```{r setup}
library(tidyverse)
library(leaflet)
```


```{r}
# download.file("https://gist.githubusercontent.com/dlebauer/66686e60bac92535585ddf7a0dfaeae2/raw/fb3f9c9a625f4a9e79221fe36906e3b7afae5f8c/cohort_met_files.csv", "data_raw/cohort_met_files.csv")

sites <- read_csv("data_raw/cohort_met_files.csv")
mandifore_raw <- 
  sites %>%
  filter(str_detect(sitename, "MANDIFORE")) #only copied MANDIFORE sites
colnames(mandifore_raw)
```

- `site_id`: BETTY site ID.  E.g. http://welsch.cyverse.org:8000/bety/sites/1000001024
- `sitename`: Long name of site
- `lon`, `lat`: location
- `format_name`: whether the path is to the met drivers or to the cohort files
- `file_path`: the path it was copied from. 
- `filename`: On Welsch these are at /data/input/...
- `start_date`, `end_date`, time range in dmy_hms format
- `duration`: end_date - start_date?
- `created_at`: when the file was created?

We don't need all columns to generate pecan.xml files.

```{r}
mandifore <- 
  mandifore_raw %>% 
  dplyr::select(-duration, -created_at, -file_path) %>% 
  mutate(across(ends_with("_date"), lubridate::dmy_hms)) %>% 
  arrange(site_id, format_name)
mandifore
```

Only keep site_id's that have both `format_name`s

```{r}
mandifore <- 
  mandifore %>%
  group_by(site_id) %>%
  filter(n()==2)
```

Split MET and cohort data and re-combine

```{r}
met <-
  mandifore %>% 
  filter(format_name == "ed.met_driver_header files format") %>% 
  rename_with(~paste0("met_", .), c(filename, start_date, end_date)) %>% 
  dplyr::select(-format_name)

cohort <- 
  mandifore %>% 
  filter(format_name == "ED2.cohort") %>% 
  rename(cohort_filename = filename) %>% 
  dplyr::select(-format_name)

new_sites <- 
  full_join(met, cohort, by = c("site_id", "sitename", "lon", "lat")) %>% 
  ungroup()
new_sites
```

Now `start_date` and `end_date` refer to the start and end of the cohort file (which may not be important) and `met_start_date` and `met_end_date` refer to start and end of the MET data.

# Extract location

PNW = pacific north west
SEUS = south east united states

```{r}
new_sites <- new_sites %>% 
  mutate(location = str_extract(sitename, "(?<=MANDIFORE-)\\w+"), .after = sitename) 
```


# Write data


```{r}
write_csv(new_sites, "data/mandifore_sites.csv")
```


# Plot sites

Plot all the sites

```{r}
leaflet(new_sites) %>% 
  addProviderTiles(providers$Esri.WorldGrayCanvas) %>% 
  addCircleMarkers(lng = ~lon, lat = ~lat, label = new_sites$sitename, 
                   radius = 1, opacity = 1, fillOpacity = 1)
```

Plot NW transect throgh SEUS sites:

```{r}
set.seed(4444)
transect_sites <- new_sites %>% 
  filter(lon > -83 & lon < -82) %>% 
  mutate(lat_round = round(lat)) %>% 
  group_by(lat_round) %>% 
  slice_sample(n=1)

leaflet(transect_sites) %>% 
  addProviderTiles(providers$Esri.WorldGrayCanvas) %>% 
  addCircleMarkers(lng = ~lon, lat = ~lat, label = new_sites$sitename, 
                   radius = 1, opacity = 1, fillOpacity = 1)
```
